candlestick chart
stock candlestick chart：This update includes the candlestick chart data set, tableau files and image format files.
Data set explanation: The highest price is the highest daily transaction price, and the lowest price is the opposite. Opening price is the daily opening price, and Closing Price is the daily closing price. Trading volume is the daily trading volume, and turnover is the transaction amount. These two data are not used in the chart, but I think they may be used in the future, so I summarized them first.